Voter turnout in Canada. Rstudio file
This research project focuses on data that is being collected through a nationally representative sample done by an online public opinion survey.  This comes from the 2021 Canadian Election Study data set. The survey consists of two parts: a campaign period survey and a post-election survey. Overall, there are 20968 individuals (who might not have participated in each survey) and there are 1062 variables. 


The aim of the research project is to determine why certain groups of Canadians have higher or lower voter turnout. In our voter turnout investigation, three graphs are created: one on who voters voted for, another on voters' education level and how it affects their support for either the Liberal or Conservative parties. Look towards the Conservative and Liberal parties of Canada, as they have been the front runners in all aspects of government. Over the last few decades, Conservatives and Liberals from either group have won the majority of votes in the House of Commons, eventually becoming our primary federal group. The last graph looks at gender and whether there is a group difference in voter participation. To be more in-depth, does one gender group have a certain preference in voting or not, and how does their region affect their choice in politics and the outcome of voter turnout?